Standardize module names in rmgpy.tools
