+++ scripts/diffModels.py
@@ -51,7 +51,7 @@ Flag                    Description
 --commonDiffOnly        Only show species and reactions present in BOTH models which have different values
 ======================= ==================================================================================== 
 """
-import rmgpy.tools.diff_models as diff_models
+import rmgpy.tools.diffmodels as diff_models
 
 
 ################################################################################
